agents/market-agent/requirements.txt
# Hyper-Sentinel v2 — MarketAgent Dependencies
# ==============================================

# Agent framework
upsonic>=0.50.0

# NATS messaging
nats-py>=2.7.0

# Data validation
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# Market data
yfinance>=0.2.30

# LLM provider SDKs
anthropic>=0.40.0
google-generativeai>=0.8.0
